requirements/scipy_mamba.txt
abnumber
anarci
altair
beautifulsoup4
black
bokeh
bottleneck
cloudpickle
conda-forge::blas=*=openblas
cython
dask
dill
h5py
ipympl
ipywidgets
matplotlib-base
modelcif
natsort
nodejs>=20.0.0
numba
numexpr
openmm
openpyxl
pandas
patsy
pdbfixer
protobuf
pyarrow
pytables
scikit-image
scikit-learn
scipy
seaborn
sqlalchemy
statsmodels
sympy
widgetsnbextension
xlrd